feat: Add development vue app
during development currently the package has to be rebuild before changes
are reflected in the examples.
To have a faster development feedback a minimal app template was added,
which can be served by running 'vite' or 'npm run dev' from the command line.

This enables consistent use of breakpoints from both within the browser or
your IDE and serves as Test for the es module format.