fix: gpx response not returned correctly
the text response from the gpx endpoint was also retrieved through
the .json function of the Response object which threw an error as it
was not possible to parse the gpx to json.

For gpx it now uses the .text function instead and just returns the
gpx result as a string.